+++ packages/orq-rc/RELEASES.md
@@ -11199,3 +11199,13 @@ Based on:
 - [typescript v4.13.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.9 - packages/orq-rc
+
+## 2026-07-28 09:45:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.10 - packages/orq-rc